Adjust styles for panels
This makes the assumption that panels are often stacked, rather than
inline.

A panel which appears between a set of radio buttons must have a margin
on the bottom of it, this is 15px (set for all panels).

Add a top margin, so that there is a gap between a form control and the
panel (but so that the two appear associated).

Add extra spacing above a panel if it is the first panel in a group,
the last panel in a group or if it is inline.

Remove any bottom margin or padding, as any content inside the panel
should sit flush with the bottom of the panel.